+++ src/graphene_elastic/filter_backends/ordering/__init__.py
@@ -1,0 +1,6 @@
+from .common import *
+
+__title__ = 'graphene_elastic.filter_backends.ordering'
+__author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
+__copyright__ = '2019 Artur Barseghyan'
+__license__ = 'GPL-2.0-only OR LGPL-2.1-or-later'