+++ src/graphene_elastic/filter_backends/queries.py
@@ -1,33 +1,32 @@
 import graphene
-
-from .. import constants
+from ..enums import NoValue
 
 __title__ = "graphene_elastic.filter_backends.queries"
 __author__ = "Artur Barseghyan <artur.barseghyan@gmail.com>"
 __copyright__ = "2019 Artur Barseghyan"
 __license__ = "GPL-2.0-only OR LGPL-2.1-or-later"
 __all__ = (
-    'Term',
-    'Terms',
-    'Range',
-    'Exists',
-    'Prefix',
-    'StartsWith',
-    'Wildcard',
-    'GeoDistance',
-    'Point',
-    'GeoPolygon',
-    'GeoBoundingBox',
     'Contains',
-    'In',
+    'Direction',
+    'EndsWith',
+    'Exclude',
+    'Exists',
+    'GeoBoundingBox',
+    'GeoDistance',
+    'GeoPolygon',
     'Gt',
     'Gte',
+    'In',
+    'IsNull',
     'Lt',
     'Lte',
-    'EndsWith',
-    'IsNull',
-    'Exclude',
-    'LOOKUP_FILTER_MAPPING',
+    'Point',
+    'Prefix',
+    'Range',
+    'StartsWith',
+    'Term',
+    'Terms',
+    'Wildcard',
 )
 
 
@@ -42,6 +41,10 @@ class _ListOfTypeString(graphene.List):
             *args,
             **kwargs
         )
+
+# ***************************************************************
+# ************************* Filtering ***************************
+# ***************************************************************
 
 
 class Term(graphene.String):
@@ -253,28 +256,14 @@ class Exclude(_ListOfTypeString):
         ]
     """
 
+# ***************************************************************
+# ************************** Ordering ***************************
+# ***************************************************************
 
-LOOKUP_FILTER_MAPPING = {
-    constants.LOOKUP_FILTER_TERM: Term,
-    constants.LOOKUP_FILTER_TERMS: Terms,
-    constants.LOOKUP_FILTER_RANGE: Range,
-    constants.LOOKUP_FILTER_EXISTS: Exists,
-    constants.LOOKUP_FILTER_PREFIX: Prefix,
-    constants.LOOKUP_FILTER_WILDCARD: Wildcard,
-    # constants.LOOKUP_FILTER_REGEXP: Regexp,
-    # constants.LOOKUP_FILTER_FUZZY: Fuzzy,
-    # constants.LOOKUP_FILTER_TYPE: Type,
-    constants.LOOKUP_FILTER_GEO_DISTANCE: GeoDistance,
-    constants.LOOKUP_FILTER_GEO_POLYGON: GeoPolygon,
-    constants.LOOKUP_FILTER_GEO_BOUNDING_BOX: GeoBoundingBox,
-    constants.LOOKUP_QUERY_CONTAINS: Contains,
-    constants.LOOKUP_QUERY_IN: In,
-    constants.LOOKUP_QUERY_GT: Gt,
-    constants.LOOKUP_QUERY_GTE: Gte,
-    constants.LOOKUP_QUERY_LT: Lt,
-    constants.LOOKUP_QUERY_LTE: Lte,
-    constants.LOOKUP_QUERY_STARTSWITH: StartsWith,
-    constants.LOOKUP_QUERY_ENDSWITH: EndsWith,
-    constants.LOOKUP_QUERY_ISNULL: IsNull,
-    constants.LOOKUP_QUERY_EXCLUDE: Exclude,
-}
+
+@graphene.Enum.from_enum
+class Direction(NoValue):
+    """Ordering direction."""
+
+    ASC = 'asc'
+    DESC = 'desc'