+++ src/graphene_elastic/tests/__init__.py
@@ -1,2 +1,0 @@
-from graphene.test import Client
-